Module nmea_attack::nmea_msg

Contains a struct to represent a NMEA message and functions to convert between NMEAMsg and strings

Messages are obtained from MSG_PTR which points to a buffer of u8s. The u8 values are unicode for the hexadecimal values that make up the message. Each unicode value represents a nibble of data, as we can only represent 0x0 to 0xF. Some values such as PGN and data_length_bytes are too large to fit within a single nibble so data is transmit most significant nibble first.

Messages are received from the buffer in the following format

buffer[0] = controller number
buffer[1] = priority
buffer[2] = most significant nibble of PGN
buffer[3] = 2nd nibble of PGN
buffer[4] = 3rd nibble of PGN
buffer[5] = 4th nibble of PGN
buffer[6] = least significan nibble of PGN
buffer[7] = source
buffer[8] = most significant nibble of data_length in bytes
buffer[9] = least significant nibble of data_length in bytes
buffer[10] and beyond = data 

Note that buffer[10] and buffer[11] would form the first byte of data

Example Message Conversion

An example message obtained from the buffer might look like this:

0x30 0x32 0x32 0x46 0x31 0x31 0x32 0x33 0x31 0x38 0x46 0x46 0x46 0x45 0x46 0x46 0x46 0x46 0x37 0x46 0x46 0x46 0x37 0x46 0x46 0x43

Those unicode values represent the following numbers:

0x0 0x2 0x1 0xF 0x1 0x1 0x2 0x3 0x0 0x8 0xF 0xF 0xF 0xE 0xF 0xF 0xF 0xF 0x7 0xF 0xF 0xF 0x7 0xF 0xF 0xC

Which then translates to:

controller number = 0x00

priority = 0x02

PGN = 0x01F112

source = 0x03

data_length_bytes = 0x08

data = [0xFF, 0xFE, 0xFF, 0xFF, 0x7F, 0xFF, 0x7F, 0xFC]

Structs

NMEAMsg

Represents a NMEA 2000 Message

Constants

ID_LENGTH 🔒

MAX_DATA_LENGTH_BYTES

Functions

PrintInt32⚠

PrintStr⚠

chars_to_nmea

Converts a char array to a NMEA message

unicode_convert 🔒

Converts a unicode number to the actual value it represents.